Make set password page UI more similar to sign in page UI
