+++ src/libs/Navigation/AppNavigator/PublicScreens.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {createStackNavigator} from '@react-navigation/stack';
 import SignInPage from '../../../pages/signin/SignInPage';
-import SetPasswordPage from '../../../pages/SetPasswordPage';
+import SetPasswordPage from '../../../pages/setpassword/SetPasswordPage';
 
 const RootStack = createStackNavigator();
 const defaultScreenOptions = {
