+++ src/pages/setpassword/SetPasswordForm/index.native.js
@@ -1,0 +1,3 @@
+import SetPasswordFormNarrow from './SetPasswordFormNarrow';
+
+export default SetPasswordFormNarrow;
